+++ responses.go
@@ -230,7 +230,7 @@ type GetTableRowsRequest struct {
 	Index      string `json:"index_position,omitempty"` // Index number, 1 - primary (first), 2 - secondary index (in order defined by multi_index), 3 - third index, etc. Number or name of index can be specified, e.g. 'secondary' or '2'.
 	EncodeType string `json:"encode_type,omitempty"`    // The encoding type of key_type (i64 , i128 , float64, float128) only support decimal encoding e.g. 'dec'" "i256 - supports both 'dec' and 'hex', ripemd160 and sha256 is 'hex' only
 	Reverse    bool   `json:"reverse,omitempty"`        // Get rows in reverse of the index
-	JSON       bool   `json:"json"`                     // JSON output if true, binary if false
+	JSON       bool   `json:"json"`                     // JSON expectOutput if true, binary if false
 }
 
 type GetTableRowsResp struct {
@@ -307,7 +307,7 @@ type GetRequiredKeysResp struct {
 }
 
 // PushTransactionFullResp unwraps the responses from a successful `push_transaction`.
-// FIXME: REVIEW the actual output, things have moved here.
+// FIXME: REVIEW the actual expectOutput, things have moved here.
 type PushTransactionFullResp struct {
 	StatusCode    string
 	TransactionID string               `json:"transaction_id"`